requirements.txt
streamlit
opencv-python
tensorflow
mediapipe
streamlit-webrtc
scikit-learn
matplotlib
seaborn
pillow